Alt-ergo

The Alt-ergo prover

The alt-ergo package installs the alt-ergo binary, whose documentation is available through the --help option.

See also the language documentation.

Alt_Ergo_common

This package uses the Alt-Ergo_common internal lib (see Alt_ergo_common) for parsing command line, input file, and main solving loop.

See also the list of modules.

Plugins

Why3 parser

For the Why3 parser plugin documentation see ABWhy3.

Fourier Motzkin simplex

For the Fourier Motzkin simplex plugin documentation see FMsimplex.