Module Compat.In_channel

Input channels.

This module provides functions for working with input channels.

See the example section below.

Channels

type t = Stdlib.in_channel

The type of input channel.

val isatty : t -> bool

isatty ic is true if ic refers to a terminal or console window, false otherwise.