Module Compat.Seq

val uncons : 'a Stdlib.Seq.t -> ('a * 'a Stdlib.Seq.t) option

If xs is empty, then uncons xs is None.

If xs is nonempty, then uncons xs is Some (x, ys) where x is the head of the sequence and ys its tail.

val is_empty : 'a Stdlib.Seq.t -> bool

is_empty xs determines whether the sequence xs is empty.

It is recommended that the sequence xs be persistent. Indeed, is_empty xs demands the head of the sequence xs, so, if xs is ephemeral, it may be the case that xs cannot be used any more after this call has taken place.

val append : 'a Stdlib.Seq.t -> 'a Stdlib.Seq.t -> 'a Stdlib.Seq.t

append xs ys is the concatenation of the sequences xs and ys.

Its elements are the elements of xs, followed by the elements of ys.

val equal : ('a -> 'b -> bool) -> 'a Stdlib.Seq.t -> 'b Stdlib.Seq.t -> bool

Provided the function eq defines an equality on elements, equal eq xs ys determines whether the sequences xs and ys are pointwise equal.

At least one of the sequences xs and ys must be finite.