Module Type.Id

Type identifiers.

A type identifier is a value that denotes a type. Given two type identifiers, they can be tested for equality to prove they denote the same type. Note that:

Unequal identifiers do not imply unequal types: a given type can be denoted by more than one identifier.

Type identifiers can be marshalled, but they get a new, distinct, identity on unmarshalling, so the equalities are lost.

Type identifiers

type 'a t

The type for identifiers for type 'a.

val make : unit -> 'a t

make () is a new type identifier.

val uid : 'a t -> int

uid id is a runtime unique identifier for id.

val provably_equal : 'a t -> 'b t -> ('a, 'b) eq option

provably_equal i0 i1 is Some Equal if identifier i0 is equal to i1 and None otherwise.