Module EphemeralDomainMap.Entry

type t

A mutable entry associated with a given key. Can be used to access and update the associated domain imperatively. A single (physical) entry is associated with a given key.

type domain

The type of domains associated with an entry.

val domain : t -> domain

Return the domain associated with this entry.

val set_domain : t -> domain -> unit

set_domain e d sets the domain of entry e to d. This overwrites any pre-existing domain associated with e.

Note: if you need to tighten an existing domain, this must be done explicitely by accessing the current domain through domain before calling set_domain. See MakEntryNotation.