Module Ephemeral.Entry

type t

A mutable entry associated with a given key. Can be used to access and update the associated domain imperatively. A single (physical) entry is associated with a given key.

type domain = domain

The type of domains associated with an entry.

val domain : t -> domain

Return the domain associated with this entry.

val set_domain : t -> domain -> unit

set_domain e d sets the domain of entry e to d. This overwrites any pre-existing domain associated with e.

Note: the caller is responsible for ensuring that the domain is a subset of the possible domains for the entry (e.g. due to type constraints). The recommended way to do so is by first intersecting with the existing domain. See also the EntryNotation functor which does this for you.