Module Options.Output

Output channels manager.

type t = private

| Stdout

| Stderr

| Channel of string * Stdlib.out_channel * Stdlib.Format.formatter

| Fmt of Stdlib.Format.formatter

| Invalid

val to_string : t -> string

to_string Returns a string representation of the output channel.

val of_formatter : Stdlib.Format.formatter -> t

of_formatter fmt create an out channel of the formatter fmt.

val to_formatter : t -> Stdlib.Format.formatter

to_formatter fmt return the underlying formatter.

val create_channel : string -> t

create_filename filename create an out channel to the file filename. If the argument is "stdout", respectively "stderr", the channel is the standard output, respectively the standard error. If the file does not exist, the procedure creates it. An existant file is truncated to zero length.

val close_all : unit -> unit

Flushing and closing all the remaining output channels.

val set_regular : t -> unit

Set the regular output channel used by default to output results, models and unsat cores.

Default to Format.std_formatter.

val set_diagnostic : t -> unit

Set the diagnostic output channel used by default to output errors, debug and warning informations.

Default to Format.err_formatter.

val set_dump_models : t -> unit

Set the models output channel used by the option `--dump-models`.

Default to Format.err_formatter.

val get_fmt_regular : unit -> Stdlib.Format.formatter

Value specifying the formatter used to output results.

Default to Format.std_formatter.

val get_fmt_diagnostic : unit -> Stdlib.Format.formatter

Value specifying the formatter used to output errors.

Default to Format.err_formatter.

val get_fmt_models : unit -> Stdlib.Format.formatter

Value specifying the formatter used to output models printed by the `--dump-models` option.

Default to Format.err_formatter.