Module AltErgoLib.Rel_utils

module X = Shostak.Combine

module MX = Shostak.MXH

module SX = Shostak.SXH

module HX = Shostak.HX

module L = Xliteral

module LR = AltErgoLib.Uf.LX

module SR : sig ... end

val assume_nontrivial_eqs : X.r Sig_rel.input list -> X.r Sig_rel.input list -> X.r Sig_rel.fact list

assume_nontrivial_eqs eqs la can be used by theories to remove from the equations eqs both duplicates and those that are implied by the assumptions in la.

type delayed_fn = Uf.t -> Symbols.operator -> Expr.t list -> (X.r * Explanation.t) option

val delay1 : (Uf.r -> 'a) -> ('b -> 'c) -> ('d -> 'e -> 'f option) -> Uf.t -> 'g -> Expr.t list -> ('h * Explanation.t) option

val delay2 : (Uf.r -> 'a) -> ('b -> 'c) -> ('d -> 'e -> 'f -> 'g option) -> Uf.t -> 'h -> Expr.t list -> ('i * Explanation.t) option

module Delayed : sig ... end

The Delayed module can be used by relations that deal with partially interpreted functions. It will introduce the equalities between a function and its interpreted value as soon as the value of its arguments become known.

module type Domain = sig ... end

module type Domains = sig ... end

Extended signature for global domains.

module Domains_make (Domain : Domain) : Domains with type elt = Domain.t

Extended signature for global domains.

type 'c acts = {

acts_add_lit_view : X.r L.view -> unit;
(*
Assert a semantic literal.
*)

acts_add_eq : X.r -> X.r -> unit;
(*
Assert equality between two semantic values.
*)

acts_add_constraint : 'c -> unit;
(*
Assert a new constraint.
*)

}

The 'c acts type is used to register new facts and constraints in Constraint.simplify.

module type Constraint = sig ... end

type 'a explained = {

value : 'a;

explanation : Explanation.t;

}

val explained : ex:Explanation.t -> 'a -> 'b explained

module Constraints_make (Constraint : Constraint) : sig ... end