Module Rel_utils.Constraints_make

Parameters

module Constraint : Constraint

Signature

type t

The type of constraint sets. A constraint set records a set of constraints that applies to semantic values, and remembers the relation between constraints and semantic values.

The constraints applying to a given semantic value can be recovered using the iter_pending functions.

New constraints are marked as "pending" when added to the constraint set (whether by a call to add or following a substitution). These constraints should ultimately be propagated; they can be accessed through the iter_pending. Once pending constraints have been propagated, the "pending" constraints should be cleared with clear_pending.

val pp : t Fmt.t

Pretty-printer for constraint sets.

val empty : t

The empty constraint set.

val add : ex:Explanation.t -> Constraint.t -> t -> t

add ~ex c t adds the constraint c to the set t.

The explanation ex justifies that the constraint c holds. If the same constraint is added multiple times with different explanations, only one of the explanations for the constraint will be kept.

val subst : ex:Explanation.t -> X.r -> X.r -> t -> t

subst ~ex p v t replaces all instances of p with v in the constraints.

The explanation ex justifies the equality p = v.

val iter_parents : (Constraint.t explained -> unit) -> X.r -> t -> unit

iter_parents f r t calls f on all the constraints that apply directly to r (precisely, all the constraints r is an argument of).

val iter_pending : (Constraint.t explained -> unit) -> t -> unit

iter_pending f t calls f on all the constraints currently marked as pending. Constraints are marked as pending when they are added, including when a new constraint is added due to substitution of an old constraint (whether the old constraint was pending or not).

val clear_pending : t -> t

clear_pending t returns a copy of t except that no constraints are marked as pending.

val has_pending : t -> bool

has_pending t returns true if there is any constraint marked as pending. Hence if has_pending t returns false, iter_pending and clear_pending are guaranteed to be no-ops. Should only be used for optimization.

val fold_args : (X.r -> 'a -> 'a) -> t -> 'a -> 'a

fold_args f t acc folds f over all the term representatives that are arguments of at least one constraint.

val simplify_pending : (X.r L.view * Explanation.t) list -> t -> (X.r L.view * Explanation.t) list * t

Simplify the pending constraints. This takes as argument a list of (explained) literals, and returns a list of (explained) literals, so that constraint simplification is able to propagate new literals (typically equalities) to the UF module.