Module Rel_utils.Domains_make

Extended signature for global domains.

Parameters

module Domain : Domain

Signature

include Uf.GlobalDomain

type t

The type of global domains.

val pp : t Fmt.t

Pretty-printer for global domains.

type Uf.id +=

| Id : t Uf.id
(*
Unique identifier for this module. Used for dispatch by the union-find.

Warning: This identifier must be unique; do not re-export the Id from another module (e.g. through include).
*)

val empty : t

The empty domain.

val filter_ty : Ty.t -> bool

Limit the type of terms that this module cares about. Only substitutions of representatives for which filter_ty (type_info r) holds will be propagated to this module.

val add : Uf.r -> t -> t

add r t is called when the representative r is added to the union-find, if it has a type that matches filter_ty.

Note: unlike Relation.add, this function is called even for "dynamic" representatives added by X.make or AC normalization.

exception Inconsistent of Explanation.t

Raised by subst below when an inconsistency is found while merging the domains.

val subst : ex:Explanation.t -> Uf.r -> Uf.r -> t -> t

subst ~ex rr nrr t is called when the representatives rr and nrr are merged with explanation ex. nrr is the new representative; rr should no longer be tracked and the intersection of domains should be associated with nrr.

The explanation ex justifies the equality rr = nrr.

raises Inconsistent
if the domains for rr and nrr are incompatible.

type elt = Domain.t

The type of domains contained in the map. Each domain of type elt applies to a single semantic value.

val get : X.r -> t -> elt

get r t returns the domain currently associated with r in t.

val fold_leaves : (X.r -> elt -> 'a -> 'a) -> t -> 'a -> 'a

fold f t acc folds f over all the domains in t that are associated with leaves.

val has_changed : t -> bool

Returns true if any element is marked as changed. This can be used to avoid unnecessary calls to edit.

Elements are marked as changed when their domain shrinks due to a call to subst, or through the ephemeral API. Elements can be unmarked by clear_changed in the ephemeral API.

module Ephemeral : sig ... end

val edit : t -> Ephemeral.t

edit d returns an ephemeral version of the domain that can be used for editing.

val snapshot : Ephemeral.t -> t

snapshot e returns a persistent version of e.