Module Domains_make.Ephemeral

type handle

A mutable handle to the domain associated with a semantic value. Can be used to access and update the domain.

val (!!) : handle -> elt

Return the domain associated with the handle.

val update : ex:Explanation.t -> handle -> elt -> unit

Intersect the domain associated with the handle with the provided domain. The explanation ex justifies that the domain applies to the handle's representative.

If this changes the domain associated with the handle, the handle is marked as changed.

raises Domain.Inconsistent
if the intersection is empty.

type t

Mutable mappings from semantic values to domains.

val handle : t -> X.r -> handle

handle t r returns the handle associated with r.

There is a unique handle associated with each semantic value r that is created on-the-fly when handle t r is called for the first time.

The domain associated with the handle is initialized from the underlying persistent domain the first time it is accessed, and updated with update.

val structural_propagation : t -> X.r -> unit

Perform structural propagation for the given representative.

More precisely, if r is a leaf, the domain of r is propagated to any semantic value that contains r as a leaf according to the structure of that semantic value (using Domain.map_leaves); if r is not a leaf, its domain is propagated to any of the leaves it contains.

We only perform *forward* structural propagation: if structural propagation causes a domain of a leaf or parent to be changed, then we only mark that leaf or parent as changed.

raises Inconsistent
if an inconsistency if detected during structural propagation.

val iter_changed : (X.r -> unit) -> t -> unit

Iterate over all the semantic values that have been marked as changed since the last call to clear_changed. Values are marked as changed by update whenever their domain shrinks.

Warning: The behavior is not specified if the ephemeral domain is modified during iteration, such as by calling update or structural_propagation.

val clear_changed : t -> unit

Remove the changed flag from all values.