Module type Rel_utils.Constraint

type t

The type of constraints.

Constraints apply to semantic values of type X.r as arguments.

val pp : t Fmt.t

Pretty-printer for constraints.

val compare : t -> t -> int

Comparison function for constraints. The comparison function is arbitrary and has no semantic meaning. You should not depend on any of its properties, other than it defines an (arbitrary) total order on constraint representations.

val fold_args : (X.r -> 'a -> 'a) -> t -> 'a -> 'a

fold_args f c acc folds function f over the arguments of constraint c.

During propagation, the constraint MUST only look at (and update) the domains associated of its arguments; it is not allowed to look at the domains of other semantic values. This allows efficient updates of the pending constraints.

val subst : X.r -> X.r -> t -> t

subst p v cs replaces all the instances of p with v in the constraint.

Substitution can perform constraint simplification.

val simplify : t -> t acts -> bool

simplify c acts simplifies the constraint c by calling appropriate functions on acts.

Note: All the facts and constraints added through acts must be logically implied by c only. Doing otherwise is a soundness bug.

Returns true if the constraint has been fully simplified and can be removed, and false otherwise.

Note: Returning true will cause the constraint to be removed, even if it was re-added with acts_add_constraint. If you want to add new facts/constraints but keep the existing constraint (usually a bad idea), return false instead.