Module type Rel_utils.Domain

type t

The type of domains for a single value.

This is an abstract type that is instanciated by the theory. Note that it is expected that this type can carry explanations.

val equal : t -> t -> bool

equal d1 d2 returns true if the domains d1 and d2 are identical. Explanations should not be taken into consideration, i.e. two domains with different explanations but identical semantics content should compare equal.

val pp : t Fmt.t

Pretty-printer for domains.

exception Inconsistent of Explanation.t

Exception raised by intersect when an inconsistency is detected.

val filter_ty : Ty.t -> bool

Filter for the types of values this domain can be attached to.

val unknown : Ty.t -> t

unknown ty returns a full domain for values of type t.

raises Invalid_argument
if filter_ty ty does not hold.

val add_explanation : ex:Explanation.t -> t -> t

add_explanation ~ex d adds the justification ex to the domain d. The returned domain is identical to the domain of d, only the justifications are changed.

val intersect : t -> t -> t

intersect d1 d2 returns a new domain d that subsumes both d1 and d2. Any explanation justifying that d1 and d2 apply to the same value must have been added to d1 and d2.

raises Inconsistent
if d1 and d2 are not compatible (the intersection would be empty).

val fold_leaves : (X.r -> t -> 'a -> 'a) -> X.r -> t -> 'a -> 'a

fold_leaves f r t acc folds f over the leaves of r, deconstructing the domain t according to the structure of r.

It is assumed that t already contains any justification required for it to apply to r.

raises Inconsistent
if r cannot possibly be in the domain of t.

val map_leaves : (X.r -> t) -> X.r -> t

map_leaves f r is the "inverse" of fold_leaves in the sense that it rebuilds a domain for r by using f to access the domain for each of r's leaves.